refactor(engine): move exception handler classes to tenant/runtime
Move ViaductDataFetcherExceptionHandler and UnwrapExceptionUtil from
engine/runtime to tenant/runtime to break the engine/runtime -> tenant/api
dependency cycle.

This enables moving CheckerDispatcher into engine/runtime without creating:
engine/api -> engine/runtime -> tenant/api -> engine/api

Changes:
- Move UnwrapExceptionUtil to tenant/runtime/execution
- Move ViaductDataFetcherExceptionHandler to tenant/runtime/execution
- Move ViaductDataFetcherExceptionHandlerTest to tenant/runtime
- Update dependencies in build.gradle.kts files:
  - engine/runtime: remove tenant/api, add tenant/runtime (test)
  - tenant/runtime: add engine/runtime, snipped/errors
  - engine/wiring: add tenant/runtime
  - service/runtime: add tenant/runtime
- Update imports in consuming files
