+++ engine/wiring/build.gradle.kts
@@ -12,6 +12,7 @@ dependencies {
     implementation(libs.viaduct.engine.api)
     implementation(libs.viaduct.engine.runtime)
     implementation(libs.viaduct.service.api)
+    implementation(libs.viaduct.tenant.runtime)
     implementation(libs.viaduct.shared.deferred)
     implementation(libs.viaduct.shared.utils)
