+++ service/runtime/build.gradle.kts
@@ -18,6 +18,7 @@ dependencies {
     implementation(libs.viaduct.engine.api)
     implementation(libs.viaduct.engine.runtime)
     implementation(libs.viaduct.engine.wiring)
+    implementation(libs.viaduct.tenant.runtime)
 
     implementation(libs.viaduct.shared.graphql)
     implementation(libs.viaduct.shared.utils)
