+++ tenant/runtime/build.gradle.kts
@@ -17,7 +17,9 @@ dependencies {
     implementation(libs.viaduct.tenant.api)
 
     implementation(libs.viaduct.engine.api)
+    implementation(libs.viaduct.engine.runtime)
     implementation(libs.viaduct.service.api)
+    implementation(libs.viaduct.snipped.errors)
 
     implementation(libs.viaduct.shared.graphql)
     implementation(libs.viaduct.shared.utils)
